feat: add goBackButton
